properly trace paramless `execute` SQL call (#135)
The Python DB API doesn't seem to work, at least with Django and
Postgres, if you call `execute(sql, ())`. This stack traces trying to
index into the empty param tuple. The signature of the function,
according to PEP 249, is

 .execute ( operation [, parameters ])

so if there are no parameters, execute should be called without that
variable.