introduce `IntervalTimer` and use it instead of `threading.Timer` (#453)
`threading.Timer` only runs the function once, and then stops the thread.
To get an interval timer, we used to start a new thread when the original
thread expired. This is wasteful, and in the worst case can lead to
runoff creation of timer threads (#450).

Instead, the `IntervalTimer` will run in an endless loop, executing the
given function every n seconds.

closes #453