Implement Trace Context standard (#660)
* added support for traceparent header, and maintain support for legacy header

On incoming requests, we first check for the presence of `traceparent`,
and fall back to `elastic-apm-traceparent`.

For outgoing requests, `traceparent` is always set. `elastic-apm-traceparent`
is set if the config option `use_elastic_traceparent_header` is set to
True. For now, this option defaults to True, but will be changed to False
after a grace period.

* also use traceparent for opentracing bridge

* added support for tracestate header propagation

Initially, we only forward the header, and don't try to parse its content.
This will only be necessary once we start adding our own data to
tracestate

* added docs for new setting

* added changelog entry