omit sync=true in default span serialization (#854)
So far, we always set `sync` to either `True` or `False`. Since Kibana 7.7,
`sync=True` spans are labeled with a "blocking" tag. For mostly synchronous
code, that label isn't helpful and might be confusing.

We now omit `sync` by default and only set it to `False` for `async_capture_span`.

A further improvement might be to detect if we run in an asynchronous framework,
and set `sync=true` for `capture_span` there.