Better error handling for AWS token fetching (#884)
* Better error handling for AWS token fetching

It turns out for docker containers on AWS elastic beanstalk, the PUT for the
token fails, and fails slowly. This is not the case for the ec2 host on which
the docker container is running. This was causing issues, compounded by the
fact that I didn't have `retries=False` on that call.

* Add changelog

* Add comments about IMDSv2