aiohttp exceptions can be normal responses (#1384)
* aiohttp exceptions can be normal responses

In aiohttp HTTPExceptions inherit from Response and can have a status
that is not 500. It even has a HTTPOk exception with a 200 status code
that can be raised.

Capturing these exception as errors in APM creates a falls view of
response and service health.

This changes the aiohttp middleware to handle HTTPExceptions different
from other exceptions. It only captures HTTPExceptions equal or above a
500 status and handles all others like a success.

* silence pyflakes failure F821

* Remove unused import

* CHANGELOG

Co-authored-by: Colton Myers <colton.myers@gmail.com>